Cleaning up the types of 'ndim' and a few other things

+++ include/dtype.pxd
@@ -68,11 +68,11 @@ cdef extern from "dynd/dtype.hpp" namespace "dynd":
         dtype get_canonical_dtype()
         type_id_t type_id()
         dtype_kind_t get_kind()
-        int get_alignment()
-        uintptr_t get_data_size()
+        size_t get_alignment()
+        size_t get_data_size()
         extended_dtype* extended()
         string_encoding_t string_encoding() except +translate_exception
-        int get_undim()
+        size_t get_undim()
         dtype get_udtype()
 
 cdef extern from "dynd/dtype_assign.hpp" namespace "dynd":
